Feature/fixbranch

+++ src/main/java/neqsim/thermo/util/example/TestNeqSimBug.java
@@ -60,7 +60,7 @@ public class TestNeqSimBug {
             fluid.initPhysicalProperties();
 
             int k = 0;
-            fluidProperties[t][k++] = (double) fluid.getNumberOfPhases(); // Mix Number of Phases
+            fluidProperties[t][k++] = fluid.getNumberOfPhases(); // Mix Number of Phases
             fluidProperties[t][k++] = fluid.getPressure("Pa"); // Mix Pressure [Pa]
             fluidProperties[t][k++] = fluid.getTemperature("K"); // Mix Temperature [K]
             fluidProperties[t][k++] = fluid.getNumberOfMoles() * 100; // Mix Mole Percent
